+++ packages/adhere/changelog/CHANGELOG.md
@@ -1,3 +1,23 @@
+# 2.10.9
+
+***
+
+2023-08-14
+
+* **adhere-ui-revoiving**
+  - 加入Swiper的config
+* **adhere-ui-ajax**
+  - 修复出现异常的时候进行reject的操作
+
+# 2.10.8
+
+***
+
+2023-08-10
+
+* **adhere-ui-tablegridlayout**
+  - 加入方向样式
+
 # 2.10.7
 
 ***